requirements.txt
asgiref==3.8.1
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.1
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==8.9.0
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
whitenoise==6.6.0
